configure: fix detection of libcrypt
The crypt() function is searched in two locations: -lcrypt and the
standard C library.

The result of the former is stored in the LIBCRYPT 'scheme' while that
of the latter is stored in the 'SYSCRYPT' scheme.

However, the check for mandatory modules looks at the CRYPT 'scheme',
and thus concludes that crypt is missing when it was successfully found.

Fix that by also storing the result of either check in the 'CRYPT'
scheme.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>