Handle shutdown of TCP transport of a VoLTE endpoint
When transport is shutdown due to failure of TCP socket the transport,
it is removed and the socket is closed. In case of VoLTE it is kept
closed until the PJSIP stack sends a new message (INVITE or REGISTER).
This happens if the IMS closes the client TCP connection after
registration. Also this can happen, if the TCP connection times out.

The stored client addresses must be removed before a new REGISTER can
be sent. If an INVITE is sent, the client addresses are kept, so that
they are used to restore the previous (failed) transport with equal IPs
and ports.

Related: SYS#7067
Change-Id: I80f1caf43a8d812c807aec0a738898e007dd0fcc