Check for SSL support when trying to restart TLS transport
If SSL is not supported by PJSip, PJ_HAS_SSL_SOCK is not set to 1.
This happens, if SSL support is disabled by configuration. Then there
is no pjsip_tls_transport_restart() function, so the block that
restarts the TLS transport must not be compiled.

This patch resolves the issue of the undefined symbol:
	pjsip_tls_transport_restart

Change-Id: Iad3b9175399c5d5dac3e0b12d786a56b2983e91b