PJSIP: Fix crash after early transport failure
If registration is performed, the TCP transport towards IMS is created.
If the transport fails early, there is no transmit data set in the
client_state. (last_tdata is NULL) In this case pjsip_tx_data_add_ref()
must not be called with a NULL pointer.

When transport does not fail during later retry, registration passes
successfully.

Change-Id: Id426781e0455cb7e48a87932ef79153defbd4737