VoLTE: Use same local IP address for RTP as for SIP
When setting up pjsip for VoLTE, a local IP or interface can be
specified in the transport section of pjsip.conf. The SIP transactions
then use this IP or IP of this interface.

The socket for the RTP/RTCP steam will then be bound to the same local
IP address. This allows dedicated routing of SIP and RTP packets
towards the interface that is connected to the IMS.