VoLTE: Track registration state and block unregister/calls, if unregistered
Unlike regular SIP, VoLTE creates a (secure) transport during
registration process. This transport must be used for re-registration,
unregistration and calls. It is not possible to make calls or
to unregister, if registration is not successfull.

To prevent triggering of calls and unregistration, an internal states
tracks the registration status and is used to rejects them, if not
registered.

Change-Id: I464b8d8f4fef7009ec86269b4f436bd648484a4d