VoLTE: Add delay before sending UPDATE, if dedicated beare is assumed up
The endpoint can be configured to ignore dedicated bearer state and
assume that it is up all ready when receiving the "Session Progress"
result. This is useful for test setups where no dedicated bearer state
is available.

Send UPDATE too eary may result in a race condition at the IMS. The IMS
will reject the call if it receives an UPDATE message before the
dedicated bearer is actually up.

To solve it, a delay of 500ms is used before sending the UPDATE message.
It is possible to alter the delay.

Change-Id: I2adbac82adbcb64f069afcbd8e0a6774d336c2e8