VoLTE: Add hack for EVS codec
It turns out that outgoing EVS audio is not working on mobile terminated
calls. A workaround is to downgrade the EVS codec from "nb-bw" to "nb"
during 200 OK (UPDATE) SDP message. It does not work, if codec is
accpedted with "nb" at the beginning. (183 Session Progress message)

Also this problem seems to occurr only when IMS is transcoding.

To enable this hack and make audio work, add the following to the
endpoint section of pjsip.conf:

hack_evs=yes

Change-Id: I86e484c69d826fb06f403f5706c4aec5cdf9fa84