VoLTE: Fix handling of broken TCP transport
If TCP connection is closed, initial registration is triggered again, to
recover secure TCP transport. Outbound calls are rejected until
registration is complete.

The TCP connection may timeout due to network problems or gets closed by
the remote end (P-CSCF). This patch will triggers immidiate registration
when this happens. If the registration fails, it is retried according to
the delay and counters in pjsip.conf.

Related: SYS#7384
Change-Id: If23c95c328d02792269984886703b1a0f7b0a6ef