VoLTE: Support configurable contact_user via pjsip.conf
In certain scenarios—especially with VoWiFi—it is necessary to set a
specific Contact-User header instead of relying on the default UUID.
This commit adds support for explicitly defining the `contact_user`
parameter in both the registration and endpoint sections of a VoLTE
endpoint in pjsip.conf.

If `contact_user` are not set, the UUID will be used.

Related: OS#6757
Change-Id: I44ed751c0aa103ba963fb0d1d296b420a3e4b7cd