Correctly clean up IPSec transport when destroying VoLTE registration
IPSec policies and states are configured in the Kernel during VoLTE
registration process. Upon a new registration, old policies and states
are rmoved.

This patch ensures that old policies and states are removed when
the VoLTE registration instance is destroyed. This happens upon change
in configuration (module reload) or if Asterisk is exitted gracefully.

There can be states that don't get removed. (src 0.0.0.0 dst 0.0.0.0)
They are created during SPI reservation and will be removed by the
kernel automatically after several minutes.

Related: SYS#7589
Change-Id: I13f62c68f8d6eafca0a1f29b12ac806e4b7381d8