VoLTE: Add IMEI to Contact header of INVITE message
The IMEI in Contact header is added to the INVITE message, if specified
in VoLTE endpoint section of pjsip.config:

imei=00000000-000000-0

This IMEI is then added to the Contact header:

+sip.instance="<urn:gsma:imei:00000000-000000-0>"

The IMEI is included as given in the config file. This means that the
IMEI must be valid and have the included dashes as shown above.

Change-Id: I638d4089db0fd2a218d338a34328b07fc693246d