VoLTE: Do not perform precondition when it is not required
If a call response is received, the "Require" header indicates whether
precondtion must be used for mobilie originating calls.

If a call INVITE is received, the "Supported" header indicates whether
preconditon must be used for mobile terminating calls.

Do not perform UPDATE procedure when precondition is not required or
supported. Even if there is no precondition, the state of the dedicated
bearer needs to be updated, so that the dial plan will continue when
executing WaitForPrecondition() during mobile terminating calls.

We want to block '183 Session Progress' only if it belongs to the
precondtion process. Other or subsequent '183 Session Progress' will be
forwarded to the call processing core, which indicate early media.

Related: SYS#7686, SYS#7682
Change-Id: Iffb9ff1e41e3fee2923b77d6a53fa938134e3331