VoLTE: Add support for forwarded mobile originating calls
When a call is forwarded from B side to C side, a new "183 Session
Progress" response is received from C side at A side. This new response
has different "tag" value in the "To:" header.

If this is the case, the precondition state machine and current QoS
parameters are reset to the state when sending the initial "INVITE"
message. The precondition process is then performed with the C side, if
required.

The new "tag" is replaced in the internal invite dialog state of PJSip.
This is required, so that the PRACK message and all other messages and
responses will use the new "tag".

Related: SYS#7530
Change-Id: I05c68a373cb03d81e3ef0b5f52d1d3a49ff69e53