VoLTE: Add transport reference to resolver section in pjsip.conf
When a new INVITE is sent, the resolver is called. It can be used to
translate the IMS name into IP address, instead of resolving it via DNS.

The secure IMS server port has been assigned during registration and is
stored in the transport state. The transport section of the VoLTE
transport must be specified for the IMS resolver, in order to resolve
the current server port.

Example of the resolver section in pjsip.conf:

  [ims.mnc001.mcc001.3gppnetwork.org]
  type=resolve
  ip=123.45.67.89
  transport=volte_ims

Use "transport=<volte transport section name>" to refer to transport
section of the VoLTE configuration.

Depends: pjproject Ied5fef684310b211696e6cac8d0972529e676abd
Related: SYS#7747
Change-Id: Ic7a082056db41551d5d77c7b3d1b83a5da4f3a28